backend/requirements.txt
# Web Framework
fastapi
uvicorn

# AI & LangChain Ecosystem
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.11
langchain-chroma==0.1.4
langchain-text-splitters==0.3.3

# Vector DB
chromadb

# Database
#oracledb

# Utils
python-dotenv
pydantic
faker
pypdf